requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
html5lib==1.0.1
idna==2.9
mechanize==0.4.2
python-dotenv==0.12.0
requests==2.23.0
six==1.12.0
soupsieve==2.0
todoist-python==8.1.1
urllib3==1.25.8
webencodings==0.5.1
